requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.40.0

# Vector Database
pymilvus>=2.5.0

# PDF Parsing
magic-pdf>=0.9.0

# LLM Client
openai>=1.0.0
tenacity>=8.0.0

# Chinese NLP
jieba>=0.42.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# LlamaIndex (Agentic RAG)
llama-index>=0.10.0
llama-index-vector-stores-milvus>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# DSPy (Prompt Optimization)
dspy-ai>=2.4.0

# Optional: Gradio Frontend
# gradio>=4.0.0

# Optional: Flash Attention (requires CUDA)
# flash-attn>=2.0.0
